build/ldpd: auto-generate ldp_vty_cmds.c from ldp_vty.xml
Having ldp_vty_cmds.c around as part of the git repository was being
a major source of confusion. Since this file is auto-generated from
ldp_vty.xml, remove it from this git repository and make it be generated
on demand by adding a make target for it.

This patch adds another dependency to build Quagga, namely the LibXML
Perl module, which can be installed with this command:
$ sudo cpan XML::LibXML

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>