lib: enhance zapi_pbr_rule structure for zapi encode
The zapi_pbr_rule structure is derived from zebra_pbr_rule, and is
defined, so that a zclient will be able to encode the zebra_pbr_rule to
send ADD_RULE or DEL_RULE command.
Adding to this, an encoding function is defined, and will be used by
remote daemon to encode that message.
Note that vrf_id is also encoded in that structure. That will permit
storing where the message is coming from. This field will be VRF_DFAULT
for PBR daemon, but may be used differently by other clients ( I think
of FS BGP).

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>