bgpd: upon FS update, handle redirect VRF for cross-VRF
In the case of FS redirect VRF request, the VRF may be a VRF NETNS.
This is not possible to do vrf route leak through vrf with NETNS
backend.
For that the BGP assumes there is a cross vrf interface that uses a
specific semantic and that permits sending traffic across vrfs.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>