bgpd: upon new table instantiation, handle default route
If a new rule is identified, a new table identifier is created.
In that table, add a default route when possible. If redirect IP rule is
identified, then add a default route to that IP address.
If redirect VRF is identified, then the assumption is that the VRF
matched is reachable through a xvrf interface. Then the default route is
the address of the interface located on that VRF.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>