zebra: handle wrap script support for ipset/iptable with port
The underlying plugin script is able to configure and return ipset rules
containing port or port range.
Note that both port src and port dst is not possible. So depending on
the kind of port range ( src, or dst),the ipset is configured
accordingly. Also, the monitoring of ipset show commands is extracting
port information.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>