zebra: avoid reusing an interface from an other vrf context
There are scenarios where interfaces are created for cross vrf usage.
So the interfaces are subsequently created in the same netns, then the
latter interface is moved to the other netns. Moreover, the index across
vrf is not unique.
There is a workflow where the new interface in the other netns is
requested to be created without having to reference an old context.
If the new interface is not created, the interface search algorithm
thinks that the interface has been renamed, whereas it is not the case.

The fix consists in avoiding getting an interface from an other NETNS,
and create the interface as desired from netlink perspective.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>