zebra: revert vrf_id of ifp interface when DEL_NETLINK received
on some cases, the interface to be removed does not have the vrf_id
correctly set, while being in VRF NETNS backend. This commit is a
workaround that permits fixing this inconsistency.
This inconsistency happens also for adding an interface, but there is no
problem since the interface is being reused.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>