pbrd: Fix nearly impossible truncation
Since we are writing into the name field which is PBR_MAP_NAMELEN
size, we are expecting this to field to be at max 100 bytes.
Newer compilers understand that the %s portion may be up to
100 bytes( because of the size of the string.  The %u portion
is expected to be 10 bytes.  So in `theory` there are situations
where we might truncate.  The reality this is never going to
happen( who is going to create a nexthop group name that is
over say 30 characters? ).  As such we are expecting the
calling function to subtract 10 from the size_t l before
we pass it in to get around this new gcc fun.

Fixes: #2163
PR=59228
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
(cherry picked from commit b915220d6dd997c2de85daf4ec3f86a354447632)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>