bgpd: add missing ecommunity flowspec to display
On some cases, the ecommunity flowspec for redirect vrf is not displayed
in all cases. On top of that, display the values if ecom can no be
decoded.
Also, sub_type and type are changed from int to u_int8_t, because the
values contains match the type and sub type of extended communities.

PR=59228
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
(cherry picked from commit 485849c6001e291a7262940e5a89905ace188be1)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>