zebra: implementation of show debugging information on wrap_script
The plugin module wrap_script is being equipped with a show routine that
dumps the status of the debugging state.
Also, the vty commands used to enable or disable the wrap_script
debugging state is being renamed. Actually it was an error to use fpm
keyword. Now the word is wrap_script.

PR=58846
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Matthieu Ternisien d'Ouville <matthieu.tdo@6wind.com>