zebra: fix incorrect if deletion when an interface returns to vr0
When an interface is moved from VRx to VR0 the following netlink sequence
can occur:
1. Creation of the interface in the VR0
2. Deletion of the interface in the VRx

In this case on the reception of the deletion message the code does:
a. Call the if_delete_update function (in zebra/interface.c) to set the
moved interface on VRx as inactive and clean information relative to
this if.
b. Call the if_handle_vrf_change function (in zebra/interface.c) to move
the interface to VR0 to allow vr deletion in the future. This operation:
   - logs the following message "corruption detected -- interface with this
   name exists already in VRF 0!" as we receive previously the netlink
   creation message of this interface in VR0
   - set the vrf_id for this if to VR0
c. Call the if_netlink_check_ifp_instance_consistency function (in
zebra/if_netlink.c) in order check the ifp consistency. This operation:
   - find a VR (here VR0) where the interface is also present
   - remove the interface from the VRx but as the vr_id of the interface
   has been previously change the interface is removed on VR0 and not
   VRx

Then zebra internal interface state is inconsistent and future behaviour
is unpredictable (opeation failure or crash can occur)

To solve this issue the vrid of the interface to remove is reset to vrx
to be sure to remove the expected interface. The warning message is
still present but the zebra internal interface state is always correct.

Fixes: b53686c52a59 ("zebra: delete interface that disappeared")
PR=58846
Signed-off-by: Thibaut Collet <thibaut.collet@6wind.com>
Acked-by: Philippe Guibert <philippe.guibert@6wind.com>