zebra: fix crash on interface deletion
Way to fix conflict on interface list when an interface is moved from
VRx to default VR when message add link in default VR is received before
the message delete link in VRx is incorrect. With this implementation
interface node in VRx is deleted twice that causes a zebra crash.

Revert "zebra: fix incorrect if deletion when an interface returns to vr0"
This reverts commit c57cea42f4e7f686cd854b8690968c0be16adb1b.
PR=58846
Signed-off-by: Thibaut Collet <thibaut.collet@6wind.com>
Acked-by: Philippe Guibert <philippe.guibert@6wind.com>