zebra: fix possible corruption when interface's VR is changed
if an interface is moved from VRx to VRy and the add link on VRy netlink
message is received before the delete link on VRx message the interface
is present in list of interface present in the default VR at the lib
level.
On the reception of the add link message a new interface structure is
created and added to the list of interface managed by the VRy at the lib
level
On the reception of the remove link message the old interface is moved
in the list of interface managed by the default VR at the lib interface.
Then check of duplicate interface with the same name in the different
VRs fails as the lookup gives the first found interface, the one
belonging in the default VR, and remove operation on the default VR is
not done.

To avoid this issue test to check if an interface with the same name is
present in two different VRs is modified by checking that the found
interface is not the same as the provided one.

PR=58846
Signed-off-by: Thibaut Collet <thibaut.collet@6wind.com>
Acked-by: Philippe Guibert <philippe.guibert@6wind.com>