doc: inform the diff between config from zebra and outside
To avoid some confusions, it is precised in the documentation that
the configuration not done from zebra will not be injected in the
configuration context. As consequence, the config file will not be
impacted by underlying network context. But also, this will not be
possible for *Zebra* to attempt to modify outside networking objects.

PR=59405
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
(cherry picked from commit 4d25da4f1eaadbe96c9005fa553e14783a79c2e4)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>