zebra: change way to manage interfaces for vrf based on netns
Way to manage vrf of an interface is different is the implementation is
based on vrf-lite or netns.

With vrf-lite implementation vrf is a characteristic of the interface
(the interface is bonded on the vrf). So creation of an interface in a
specific vrf requests at least two RTM_NEWLINK netlink messages all
received in the default netns:
1. Creation of the interface. The interface belongs to the default vrf
2. Provide the vrf information through the tb[IFLA_MASTER] field. The
interface belongs now to the specified vrf

With netns implementation vrf is the id of the netns where the interface
is created. So creation of an interface in a specific vrf requests only one
RTM_NEWLINK netlink messages received in the expected vrf.
1. Creation of the interface. The interface belongs to the vrf that
sends the netlink message.

This architecture difference implies the following consequences:
- With vrf-lite the vrf-id of an interface can change that is not
possible with netns. With netns a move of an interface in a different
vrf implies a cretion message in the new vrf and a deletion in the old vrf
- With vrf-lite it is not possible to have two interfaces with the same
name in two different vrfs. The name interface is sufficient to specify
an interface whereas the couple (name, vrf-id) is needed to specify an
interface with the netns implementation.
- with vrf-lite frr interface configuration is set per interface name
whereas with netns frr interface configuration is set per tuple (interface
name, vrf id). So with vrf-lite frr interface configuration can be set
before the vrf creation whereas it is not possible with netns.

To manage this difference, in the most simple way, test on the used
architecture is needed in the following part of the code:
- if_get_by_name function (that checks if an interface exists and creates
  it if needed). With netns implementation search of the interface is
  based on the (name, vrf-id) tuple and interface must be created if
  not found. With vrf-lite search of the interface is based only on the
  name and a test on the vrf-id must be done to know if the interface is
  bonded to a new vrf.
- deletion of an interface when a RTM_DELLINK netlink message is received.
  With netns the interface is not moved to the default vrfd to avoid
  collision with interface on the default vrf that can have the same name.

This decision must be done for any vrf, even the default one. To allow any
frr daemons to know which implementation is used the vrf_is_mapped_on_netns
is updated to simply test if a netns is associated to the vrf and the
default netns is attached to the default vrf in case of netns
implementation for test on default vrf.

This simple implementation will partially or totally revert:
commit e9e9b1150f0c ("lib: create interface even if name is the same")
commit aef206aff3a5 ("zebra: improve logs when replacing interface to an
other netns")
commit b53686c52a59 ("zebra: delete interface that disappeared")
PR=59063
Signed-off-by: Thibaut Collet <thibaut.collet@6wind.com>
Acked-by: Philippe Guibert <philippe.guibert@6wind.com>