bgpd: extend enumerate API to handle or operations
The flowspec enumerate list can either be and values or or values.
In the latter case, a list is created that will be used later.
Also, the API supports the check for both and or or operations. This API
does not permit to handle both and and or operations at the same time.
The list will have to be either and or or. An other API retrieves the
operator unary value that is used: and or or. or 0 is the two operators
are used at the same time.

PR=59228
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
(cherry picked from commit 6d47669ac11b045c9adba9dcfc1161f6de7180a6)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>