doc: add information about dynamic update of default vrf name
It is possible to dynamically change default VRF name, if vrf backend is
a netns backend. By creating a link to the default netns in
/var/run/netns folder, then the file name will be used to name the
default VRF. If no backend netns is chosen, it is explained that it is
still possible to statically configure the default vrf name to new
define.

PR=59408
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
(cherry picked from commit c94e67258c220e72d9840a2a445e10099330f147)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Matthieu Ternisien d'Ouville <matthieu.tdo@6wind.com>