bgpd: fix rpki segfault
If a cache server was added after rpki was started it's tr_socket would
not be initialized. This would lead to a segfault if the rtr manager
ever decides to switch to that socket or if rpki support is stopped.

PR=59228
Signed-off-by: Marcel Röthke <marcel.roethke@haw-hamburg.de>
(cherry picked from commit 768812900b100fd1c18f8b6438f9e004b858f2db)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>