zebra: upon RTM_NEWROUTE/DELROUTE, find appropriate interface
The interface lookup algorithm is different according to if we are on
netns vrf or not. If we are on the former case, then we only have to
parse the interfaces of the netns, while if we are on the other case, we
have to parse all the interfaces of all the vrfs ( since index is not
overlapping in the latter case).

PR=59316
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
(cherry picked from commit 8da0f092884f854e5aa2b084147f9dcd1def6572)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Thibaut Collet <thibaut.collet@6wind.com>