zebra: upon new interface event, ipv6 address may need to be retrieved
There are cases where, when interface is down and is just created or
renamed, there are some ipv6 addresses. Because interface is down, ipv6
information is not passed in netlink information, and zebra has not full
awareness of the whole list of ipv6 addresses of the interface.
To comply with that, once a new interface is detected, that interface is
polled again to get ipv6 address of that interface.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>