zebra: get mpls nexthop label and append label in local route
in vrf route leak, if the entry is a routing action with a mpls add
action, that implies that the mpls switching will be applied in remote
vrf. in BGP L3VPN case, the second encapsulation is not done, because
remote nexthop IP has not knowledge of MPLS nexthop label. The MPLS
nexthop label is searched and appended in local static route entry
towards veth interface.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>