zebra: Send correct default vrf tableid for MROUTE stats
So the linux kernel uses the RT_TABLE_MAIN for the table
id used for ip routing.  The multicast routing tables use
RT_TABLE_DEFAULT.  We changed the internal code of zebra_vrf
a few months back to use RT_TABLE_MAIN as the tableid to
use.  This caused the pim sg stats to stop working because
of the kernel bug where it uses a different table
for ip routing and ip multicast.

Put a bit of a special case in to do the right thing.

PR=60818
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
(cherry picked from commit 57a1d0920d0ec0ed3fa064fe31c05b72e44207d3)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Matthieu Ternisien d'Ouville <matthieu.tdo@6wind.com>