pimd: Actually create vif's in non-integrated config
The startup of a non-integrated config was not properly
allowing for startup to create the vif when we have
not learned about the interface we are trying to configure
at this point in time.  Actually notice when we are
trying to create a pimreg device or not to properly
notice when to attempt to create the vif or not.

PR=60818
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
(cherry picked from commit 480129597d1262c36d9e4e20d29faf7fddca695e)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Matthieu Ternisien d'Ouville <matthieu.tdo@6wind.com>