lib: use an API to return the nexthop interface for a vrf.
The vrf is reachable by a nexthop interface that varies according to if
vrf is a netns based or a vrf lite. In the latter case, the if name is
the vrf name itself, while for netns based vrfs, if there are no cross
vrfs, then there is no interface to return.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>