bgpd: Fix crash when using v4 route w/ v6 nexthop
Recent changes to the nht code in bgp caused us to actually
keep a true count of v6 nexthop paths when using v4 over v6.
This change introduced a race condition on shutdown on who
got to the bnc cache first( the v4 table or not ).  Effectively
we were allowing the continued existence of the path->nexthop
pointing to the freed bnc.  This was especially true when
we had route leaking.   So when we free the bnc make sure
we clean up the path->nexthop variables pointing at it too.

PR=61154
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
(cherry picked from commit 7f040da1a18981277a1db29764a3dc6f0e7d2f10)
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>