lib: add new API to get vrf_id of the parent daemon for distrib list
distribute list are executed by daemons. So, to know on which vrf the
configuration happens, it is necessary to ask the daemon which vrf it
is. If it is not possible, then vrf default is taken. the api is
changed accordingly for running-config of distribute list, as well as
for distribute list general information.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>