lib: check if two vrf have an ipless interface
if vrf have an ipless interface in between, then the upper daemon can be
informed so that the route entry will differ slightly. this API can be
used in the following cases:
- a route leak must be installed, and one must know if it is possible to
  do vrf route leak; ie if it is a vrf-lite, then support is ok;
  however, if it is a netns based vrf, the check for veth pair presence
  is needed, as well as the operational status of the interface
- a link up/link down happened, and one must know if that event may
  impact veth pairs that can be used for crossing vrf borders. here, too
  that function can be used.

As returned information, the interface index of the interface to use may
be returned.
As passed information, either both vrf, or only the source vrf, with an
interface pointer that may be the interface of a veth pair.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>

(cherry picked from commit 89cbadf4348e)
PR=62599
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Emmanuel Vize <emmanuel.vize@6wind.com>