bgpd: ifindex and mpls labels used for route leak with nht
bnc uses a list of mpls labels that may be locally used for crossing vrf
route leak. because passing through an interface requires some specific
lsp action on default  vrf, that list will contain the association with
the originated label and the new label. That list of labels is presumably
made up of only one label for all route entries in a vrf that want to
reach a remote entry.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>