zebra: incoming impls flow from def to vrf uses cross vrf interface
in case vrf mpls add is called, a lsp entry is added.
If the cross vrf interface is present, but not down or the pair veth
interface is not present, then append however the lsp entry. the reason
for doing this is that there is not nexthop tracking facility for mpls
entries.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>