lib: add new APIs for handling vrf with vrf_id unknown
in the case the vrf name is known, but not the vrf identifier, the
respective daemon may be able to create a vrf entity. This entity will
be updated later ( this is an other subject). However, in this state,
this should be possible for the daemon to create and appen interfaces in
that vrf. For that, new APIs are defined:
- if_lookup_by_name_vrf
- if_get_by_name_vrf

those interfaces use vrf pointer instead of vrf_id.
Those APIs do the same as their respective function without _vrf suffix

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>

(cherry picked from commit 2db6567f9e2f)
PR=62623
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Emmanuel Vize <emmanuel.vize@6wind.com>