lib: define messages for vrf reachability
zebra is aware of the vrf reachability, when daemons want to create
an entry that performs a vrf route leak. for that, zebra will receive
registrations and unregistrations coming from daemons. in return, zebra
will update the information with an other identifier VRF_REACH_UPDATE.
the daemons will then benefit from this information, by using the hook
defined for being called back by ZAPI, or by using the registration ids
in the zapi interface.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>