zebra: add redistribute table-direct support on zebra code
Today, to redistribute routes from a given routing table to
a given routing protocol, the route entries have to be copied
to the main routing table with the 'ip import-table' command;
then those routes is set to a specific 'table' route type, which
is caught by the 'redistribute table' command of the routing
protocol.

The below configuration example is used:

> router bgp 65500
>  address-family ipv4 unicast
>   redistribute table 100
>  exit-address-family
> exit
> ip import-table 100

This approach is limited as the 'import-table' parameter only
handles route table up to 252. Also, duplicating the route
entries may introduce conflicts between routes in the main routing
table, and routes from the original source table.

The proposed solution consists in defining a new type of
redistribution that directly takes the route entries from a given
routing table.
- as the 'redistribute table' combination is already used, the
'redistribute table-direct' command is proposed.
- this command redistributes route entries from a given routing table
to the control plane service that requests it.
- the 'direct' keyword means that the route entries do not have to be
copied to the default routing tables, and are directly sent to the
requester.

The below configuration example is proposed:

> router bgp 65500
>  address-family ipv4-unicast
>   redistribute table-direct 100
>  exit-address-family
> exit

This first commit introduces the command in the zebra daemon, and
handles the redistribution to the control plane services: the
target control plane service must be on the default vrf instance,
and the source table must not be the default routing table

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>