bgpd: fix nht resolution for l3vpn traffic over gre interfaces
When a route imported from l3vpn is analysed, the nexthop from default
VRF is looked up against a valid MPLS path. Generally, this is done on
backbones with a MPLS signalisation transport layer like LDP. Generally,
the BGP connection is multiple hops away. That scenario is already
working.

There is case where it is possible to run L3VPN over GRE interfaces, and
where remote BGP is 1 hop away, that is to say just behind GRE interface.
In this case, the backbone is an IP backbone, and GRE is just here to
tunnel MPLS traffic. On that case, the nexthop given in the path may not
have MPLS path, but should be connected. This is where that commit is
done for, as it ensures that if the nexthop is connected, the path can
become valid, and the imported route entry can be inserted in the system.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>