(step0) isisd: update struct isis_sr_psid_info with nh context
Prefix-SID info now contains new information about nexthops
and its backup nexthops.

Before this commit, Prefix-SID info is included in
isis_route_info, and it's keeping the same nexthops as this
route-info.

However, when Flex-Algo is introduced, there will be
multiple Prefix-SIDs for a single prefix, each with a
different Algorithm.  Each of these Prefix-SIDs with
Algorithm performs SPF calculation with a separate
contract and sets a nexthops, so it is necessary to
keep a completely different nexthops for each Prefix-SID.

After this commit, the nexthops for each Prefix-SID is not
taken from route_info, but the nexthop set inside the
Prefix-SID is taken. This works for backup nexthops as well.

Signed-off-by: Hiroki Shirokura <hiroki.shirokura@linecorp.com>