zebra,*: rnh registration with colored rnh partial support
This commit addresses the case where an SR policy has been put
in place and is up, before the BGP acquires a colored update.
In those conditions, BGP creates a specific BGP nexthop entry
for the tuple (color, nexthop prefix), but is not informed of
the SR policy nexthop calculated.
To illustrate, the below 'show bgp nexthop' output shows that
the nexthop resolution for the tuple ('9.9.9.9', color 30) is
the same as for the default ('9.9.9.9', 0) tuple.

Current BGP nexthop cache:
 9.9.9.9 valid [IGP metric 50], #paths 3, peer 9.9.9.9
  gate 10.1.0.1, if eth-rt1
  gate 10.5.0.5, if eth-rt5
  Last update: Fri Oct  7 11:01:10 2022
 SR-TE color 30 - 9.9.9.9 valid [IGP metric 50], #paths 1
  gate 10.1.0.1, if eth-rt1
  gate 10.5.0.5, if eth-rt5
  Last update: Fri Oct  7 11:01:10 2022

To solve this issue, upon creating a colored Bgp nexthop context
(BNC), send a nexthop registration to zebra with the srte color.
The Zebra nexthop tracking service will read the srte color, and
will call the Zebra srte service to check if a matching sr policy
is present. The Zebra srte will send back nexthop information.

The new output is as follows:

dut-vm# show bgp nexthop
Current BGP nexthop cache:
 9.9.9.9 valid [IGP metric 50], #paths 3, peer 9.9.9.9
  gate 10.1.0.1, if eth-rt1
  gate 10.5.0.5, if eth-rt5
  Last update: Fri Oct  7 11:43:50 2022
 SR-TE color 30 - 9.9.9.9 valid [IGP metric 0], #paths 1
  gate 10.5.0.5, if eth-rt5
  Last update: Fri Oct  7 11:43:51 2022

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>