bgpd: add label information in ADJ-RIB-IN
This commit addresses the case where soft inbound
reconfiguration is applied to incoming BGP VPN prefixes.

When no operation is performed between ADJ-RIB-IN and RIB,
then the received update is the same as the update stored
in the RIB. The below illustration can help:

  > ubuntu2004# show bgp ipv4 vpn 10.200.0.0/24
  > BGP routing table entry for 444:2:10.200.0.0/24, version 4
  > not allocated
  > Paths: (1 available, best #1, accept local nexthop)
  >   Advertised to non peer-group peers:
  >   10.125.0.2
  >    65501
  >     10.125.0.2 from 10.125.0.2 (2.2.2.2)
  >       Origin incomplete, metric 0, ...
  >       Community: accept-own-nexthop
  >       Extended Community: RT:52:100
  >       Remote label: 102
  >      Last update: Thu Jan 12 11:10:46 2023

If an user performs filtering with a route-map, the incoming
update information will be stored in the adj-rib-in. The above
show command is empty, and the below show command does not
show any label information:

  > ubuntu2004# show bgp ipv4 vpn neighbors 10.125.0.2 received-routes
  > BGP table version is 4, local router ID is 1.1.1.1, vrf id 0
  > Default local pref 100, local AS 65500
  > Status codes:  s suppressed, d damped, h history, * valid, > best, = multipath,
  >                i internal, r RIB-failure, S Stale, R Removed
  > Nexthop codes: @NNN nexthop's vrf id, < announce-nh-self
  > Origin codes:  i - IGP, e - EGP, ? - incomplete
  > RPKI validation codes: V valid, I invalid, N Not found
  >
  >     Network          Next Hop            Metric LocPrf Weight Path
  > Route Distinguisher: 444:2
  >  *> 10.200.0.0/24    10.125.0.2               0     43      0 65501 ?

Assuming the user updates its route-map without flushing the BGP
adjacency, and allows the incoming update not to be filtered, then
the received update has an invalid label.

  > ubuntu2004# show bgp ipv4 vpn 10.200.0.0/24
  > [..]
  >   65501
  >     10.125.0.2 from 10.125.0.2 (2.2.2.2)
  >       Origin incomplete, metric 0, valid, external, best (First path received)
  >       Extended Community: RT:52:100
  >       Last update: Thu Jan 12 11:55:00 2023

This commit duplicates the label information in the incoming
ADJ-RIB-IN, and allows keeping the label value intact.

This commit will also be very helpful to handle BMP pre-policy
label information (out of scope of this commit).

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>